+++ src/app/scheme.rs
@@ -191,6 +191,11 @@ impl App {
                             SchemeType::Square,
                             SchemeType::Square.as_ref(),
                         );
+                        ui.selectable_value(
+                            &mut self.picker.scheme_type,
+                            SchemeType::Monochromatic,
+                            SchemeType::Monochromatic.as_ref(),
+                        );
                     });
                 ui.add(
                     Slider::new(&mut self.picker.scheme_color_size, 100.0..=250.)
@@ -263,6 +268,22 @@ impl App {
                             let c1 = s.0;
                             let c2 = s.1;
                             let c3 = s.2;
+                            ui.horizontal(|ui| {
+                                cb!(color, ui);
+                                cb!(c1, ui);
+                            });
+                            ui.horizontal(|ui| {
+                                cb!(c2, ui);
+                                cb!(c3, ui);
+                            });
+                        });
+                    }
+                    SchemeType::Monochromatic => {
+                        let mono = color.monochromatic();
+                        ui.vertical(|ui| {
+                            let c1 = mono.0;
+                            let c2 = mono.1;
+                            let c3 = mono.2;
                             ui.horizontal(|ui| {
                                 cb!(color, ui);
                                 cb!(c1, ui);