+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add settings to disable each colorspace [#13](https://github.com/vv9k/epick/pull/13)
 - Updated light theme colors
 - Add square colors
+- Add monochromatic colors
 
 # 0.4.0